backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
python-dotenv
httpx
email-validator
bcrypt==3.2.0
